copy dubbo.tag cross thread with plugin
