+++ dubbo-distribution/dubbo-all/pom.xml
@@ -430,6 +430,14 @@
             <optional>true</optional>
         </dependency>
 
+        <dependency>
+            <groupId>org.apache.dubbo</groupId>
+            <artifactId>dubbo-plugin-cross-thread</artifactId>
+            <version>${project.version}</version>
+            <scope>compile</scope>
+            <optional>true</optional>
+        </dependency>
+
         <!-- registry -->
         <dependency>
             <groupId>org.apache.dubbo</groupId>
