+++ dubbo-distribution/dubbo-bom/pom.xml
@@ -442,6 +442,11 @@
                 <version>${project.version}</version>
             </dependency>
 
+            <dependency>
+                <groupId>org.apache.dubbo</groupId>
+                <artifactId>dubbo-plugin-cross-thread</artifactId>
+                <version>${project.version}</version>
+            </dependency>
 
             <!-- registry -->
             <dependency>
