+++ dubbo-plugin/dubbo-plugin-cross-thread/src/main/java/org/apache/dubbo/crossthread/interceptor/RunnableOrCallableMethodInterceptor.java
@@ -14,25 +14,26 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.dubbo.spring.boot.toolkit;
+package org.apache.dubbo.crossthread.interceptor;
 
-import java.util.function.Consumer;
+import net.bytebuddy.asm.Advice;
 
-@DubboCrossThread
-public class ConsumerWrapper<V> implements Consumer<V> {
-    final Consumer<V> consumer;
+import org.apache.dubbo.common.constants.CommonConstants;
+import org.apache.dubbo.rpc.RpcContext;
 
-    public ConsumerWrapper(Consumer<V> consumer) {
-        this.consumer = consumer;
+public class RunnableOrCallableMethodInterceptor {
+
+    @Advice.OnMethodEnter
+    public static void onMethodEnter(
+        @Advice.FieldValue(value = RunnableOrCallableActivation.FIELD_NAME_DUBBO_TAG, readOnly = false) String dubboTag) {
+        // copy tag to RpcContext from RunnableOrCallable's field value
+        RpcContext.getClientAttachment().setAttachment(CommonConstants.TAG_KEY, dubboTag);
     }
 
-    public static <V> ConsumerWrapper<V> of(Consumer<V> consumer) {
-        return new ConsumerWrapper(consumer);
-    }
-
-    @Override
-    public void accept(V v) {
-        this.consumer.accept(v);
+    @Advice.OnMethodExit
+    public static void onMethodExit() {
+        // clear tag in RpcContext
+        RpcContext.getClientAttachment().removeAttachment(CommonConstants.TAG_KEY);
     }
 
 }
