+++ dubbo-plugin/pom.xml
@@ -50,6 +50,7 @@
         <module>dubbo-plugin-classloader-filter</module>
         <module>dubbo-plugin-proxy-bytebuddy</module>
         <module>dubbo-plugin-qos-trace</module>
+        <module>dubbo-plugin-cross-thread</module>
     </modules>
     <properties>
         <skip_maven_deploy>false</skip_maven_deploy>
