+++ dubbo-spring-boot/dubbo-spring-boot-interceptor/README.md
@@ -1,6 +1,6 @@
 # Dubbo Spring Boot interceptor
 
-`dubbo-spring-boot-interceptor` copy the header(dubbo-tag)/cookie(dubbo.tag)/urlParameter(dubbo.tag) to dubbo . 
+`dubbo-spring-boot-interceptor` copy the header(dubbo-tag)/cookie(dubbo.tag)/urlParameter(dubbo.tag) to dubbo .
 
 
 
@@ -18,8 +18,7 @@ public class WebMvcConfig implements WebMvcConfigurer {
     }
 }
 ```
-### copy the cookie(dubbo.tag=tagx) to dubbo 
-
+### copy the cookie(dubbo.tag=tagx) to dubbo
 ```
 @Configuration
 public class WebMvcConfig implements WebMvcConfigurer {
@@ -28,45 +27,4 @@ public class WebMvcConfig implements WebMvcConfigurer {
         registry.addInterceptor(new DubboTagCookieInterceptor()).addPathPatterns("/*").excludePathPatterns("/admin");
     }
 }
-```
-
-### copy dubbo.tag cross thread
-- start byte-buddy , intercept Callable or Runnable
-```
-@SpringBootApplication
-@ComponentScan(basePackages = "org.apache.your-package")
-public class SpringBootDemoApplication {
-
-    public static void main(String[] args) {
-       SpringApplication application = new SpringApplication(SpringBootDemoApplication.class);
-       application.addListeners(new DubboCrossThreadAnnotationListener(ByteBuddyAgent.install()));
-       application.run(args);
-    }
-}
-```
-- wrap Callable or Runnable
-```
-@DubboCrossThread
-public class TargetClass implements Runnable{
-    @Override
-    public void run() {
-        // ...
-    }
-}
-```
-```
-Callable<String> callable = CallableWrapper.of(new Callable<String>() {
-    @Override
-    public String call() throws Exception {
-        return null;
-    }
-});
-```
-```
-Runnable runnable = RunnableWrapper.of(new Runnable() {
-    @Override
-    public void run() {
-        // ...
-    }
-});
 ```
