+++ dubbo-test/dubbo-dependencies-all/pom.xml
@@ -365,6 +365,12 @@
             <version>${project.version}</version>
         </dependency>
 
+        <dependency>
+            <groupId>org.apache.dubbo</groupId>
+            <artifactId>dubbo-plugin-cross-thread</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+
 
         <!-- registry -->
         <dependency>
